Refactor parse functions and timetable
The returned timetable now includes ordered days which in turn include
ordered courses. The timetable also contains the associated `DegreeProgram`
for which it is valid. This instance can be nil as the server might not
return it because the given ID is not in the `Winter` semester.
For an explanation why the semester matters, see the documentation for
`FillDegreeProgram`.